+++ src/mg/http/HTTPClient.hpp
@@ -17,12 +17,17 @@ namespace be::mg {
         JSValue callback = JS_NULL ;
         struct mg_connection * conn = NULL ;
         std::string _host = ""; // 用于 tls
-        bool is_ws: 1 = false ;
-        bool is_tls:1  = false ;
-        bool is_connected = false ;
+        bool _isWS: 1 = false ;
+        bool _isTLS:1  = false ;
+        bool _isConnected = false ;
         bool _enableChunkEvent:1  = false ;
         size_t headerLength = 0 ;
         size_t receivedBodyLength = 0 ;
+
+        // client auth (双向认证)
+        bool useClientCert = false ;
+        std::string clientCert = "" ;
+        std::string clientKey = "" ;
 
         static void eventHandler(struct mg_connection * conn, int ev, void *ev_data) ;
         static void wsEventHandler(struct mg_connection * conn, int ev, void *ev_data) ;
@@ -39,6 +44,9 @@ namespace be::mg {
         static JSValue connect(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;
         static JSValue connectWS(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;
         static JSValue isConnected(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;
+        static JSValue setClientKey(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;
+        static JSValue enableClientAuth(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;
+        static JSValue disableClientAuth(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;
 
         static JSValue enableChunkEvent(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;